@@ -1,44 +1,21 @@
 class Admin::EventsController < Admin::BaseController
-  # GET /events
-  # GET /events.json
+  before_action :event, only: [:show, :edit, :destroy]
+
   def index
     @events = Event.order('id')
-
-    respond_to do |format|
-      format.html # index.html.erb
-      format.json { render json: @events }
-    end
   end
 
-  # GET /events/1
-  # GET /events/1.json
   def show
-    @event = Event.find(params[:id])
-
-    respond_to do |format|
-      format.html # show.html.erb
-      format.json { render json: @event }
-    end
   end
 
-  # GET /events/new
-  # GET /events/new.json
   def new
     @event = Event.new
-
-    respond_to do |format|
-      format.html # new.html.erb
-      format.json { render json: @event }
-    end
   end
 
-  # GET /events/1/edit
   def edit
     @event = Event.find(params[:id])
   end
 
-  # POST /events
-  # POST /events.json
   def create
     service = EventsService.new(as: current_user)
     @event = service.create(params[:event])
@@ -54,8 +31,6 @@ class Admin::EventsController < Admin::BaseController
     end
   end
 
-  # PUT /events/1
-  # PUT /events/1.json
   def update
     service = EventsService.new(as: current_user)
     @event = service.update(params[:id], params[:event])
@@ -71,10 +46,7 @@ class Admin::EventsController < Admin::BaseController
     end
   end
 
-  # DELETE /events/1
-  # DELETE /events/1.json
   def destroy
-    @event = Event.find(params[:id])
     @event.destroy
 
     respond_to do |format|
@@ -82,4 +54,9 @@ class Admin::EventsController < Admin::BaseController
       format.json { head :no_content }
     end
   end
+
+  private
+    def event
+      @event = Event.find(params[:id])
+    end
 end
