@@ -6,7 +6,7 @@ class ApplicationController < ActionController::Base
   # TODO: remove username check, generate username
   def after_sign_in_path_for(resource)
     if resource.username.present?
-      resource.sign_in_count <= 1 ? '/dashboard/user_availabilities' : root_path
+      resource.sign_in_count <= 1 ? '/dashboard/user_availabilities' : authenticated_root_path
     else
       flash[:notice] = 'Please enter your username'
       edit_profile_users_path
