@@ -1,5 +1,6 @@
 class Charity < ActiveRecord::Base
+	has_many :user_charities
+	has_many :users, through: :user_charities
 
-has_many :user_charities
-has_many :users, through: :user_charities
+	validates :name, presence: true
 end
