@@ -1,6 +1,6 @@
 class Day < ActiveRecord::Base
 
-  default_scope order('date ASC')
+  default_scope -> { order('date ASC') }
 
   has_many :user_availabilities
   has_many :department_blocks
